# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for User Service",
        "title": "API Documentation",
        "contact": {},
        "version": "0.2"
    },
    "paths": {
...
        "/numerical-method/least-squares-regression/linear-regression": {
            "post": {
                "description": "Create the linear regression result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Linear Regression"
                ],
                "summary": "Create Linear Regression Result",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqLinearRegression"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.LinearRegression"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/least-squares-regression/linear-regression/{id}": {
            "get": {
                "description": "Get the linear regression result by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Linear Regression"
                ],
                "summary": "Get Linear Regression Result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Linear Regression ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LinearRegression"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/least-squares-regression/multiple-regression": {
            "post": {
                "description": "Create the multiple regression result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Multiple Regression"
                ],
                "summary": "Create Multiple Regression Result",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqMultipleRegression"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.MultipleRegression"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/least-squares-regression/multiple-regression/{id}": {
            "get": {
                "description": "Get the multiple regression result by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Multiple Regression"
                ],
                "summary": "Get Multiple Regression Result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Multiple Regression ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MultipleRegression"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/least-squares-regression/polynomial-regression": {
            "post": {
                "description": "Create the polynomial regression result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Polynomial Regression"
                ],
                "summary": "Create Polynomial Regression Result",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqPolynomialRegression"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.PolynomialRegression"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/least-squares-regression/polynomial-regression/{id}": {
            "get": {
                "description": "Get the polynomial regression result by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Polynomial Regression"
                ],
                "summary": "Get Polynomial Regression Result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Polynomial Regression ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PolynomialRegression"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.LinearRegression": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
...
        "models.MultipleRegression": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "string"
                }
            }
        },
...
        "models.PolynomialRegression": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "order": {
                    "type": "integer"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "validations.ReqLinearRegression": {
            "type": "object",
            "properties": {
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
...
        "validations.ReqMultipleRegression": {
            "type": "object",
            "properties": {
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "string"
                }
            }
        },
...
        "validations.ReqPolynomialRegression": {
            "type": "object",
            "properties": {
                "order": {
                    "type": "integer"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
...
    }
}